+++ packages/transfer_manager_android/android/src/main/kotlin/com/pinzarulab/transfer_manager_android/TransferManagerAndroidPlugin.kt
@@ -282,6 +282,8 @@ class TransferManagerAndroidPlugin :
                 (call.argument<String>("notificationTitle") ?: "Downloading file"),
             DownloadWorker.KEY_SHOW_NOTIFICATION to
                 (call.argument<Boolean>("showNotification") ?: true),
+            DownloadWorker.KEY_NOTIFICATION_OPEN_TYPE to
+                (call.argument<String>("notificationOpenType") ?: "open"),
             DownloadWorker.KEY_MAX_ATTEMPTS to
                 (call.argument<Int>("maxAttempts") ?: 5).coerceAtLeast(1),
         )